change seed script and csvs for cassandra
